added all the Schemas and some demo DB operation
