+++ server/api/models/schemas/bidInfo.schema.js
