+++ server/api/models/schemas/eventInfo.schema.js
